refactor(python): split story.py into focused submodules
story.py (916 lines) split into: story_reader (types+constants+parsing), story_search (full-text index), story_memoir (chardict discovery), story_summary (event/chapter summaries). story.py is now a re-export shim preserving all public + test-private symbols.

No behaviour change. All 194 tests pass.
