+++ python/src/prts_mcp/data/story.py
@@ -1,915 +1,91 @@
 """Story data reader for PRTS-MCP.
 
-Reads from the bundled/synced zh_CN.zip (ArknightsStoryJson fork release).
+This module is now a compatibility re-export shim. The implementation has
+been split into focused submodules for clarity (see STYLE.md file-size
+guidelines):
 
-Zip internal layout (all paths prefixed with "zh_CN/"):
-  zh_CN/storyinfo.json                       — {story_key: summary_text}
-  zh_CN/gamedata/excel/story_review_table.json — event metadata + ordered chapter list
-  zh_CN/gamedata/story/{story_key}.json       — per-chapter dialogue JSON
+- story_reader: types, constants, and chapter/event parsing
+- story_search: full-text search index and search_stories
+- story_memoir: operator memoir discovery via chardict.json
+- story_summary: event and per-chapter summaries
+
+All public symbols are re-exported here so existing
+``from prts_mcp.data.story import ...`` imports continue to work unchanged.
 """
 from __future__ import annotations
 
-import json
-import re
-from functools import lru_cache
-from dataclasses import dataclass
-from pathlib import Path
-from typing import Literal
-
-from prts_mcp.data.stores import DirectoryStore, JsonStore, ZipStore
-
-# ---------------------------------------------------------------------------
-# Zip path constants
-# ---------------------------------------------------------------------------
-
-_STORY_REVIEW_TABLE = "zh_CN/gamedata/excel/story_review_table.json"
-_STORYINFO = "zh_CN/storyinfo.json"
-_SUMMARIES = "zh_CN/summaries.json"
-_EVENT_SUMMARIES = "zh_CN/event_summaries.json"
-_CHARDICT = "zh_CN/chardict.json"
-
-# entryType values → user-facing category strings
-_CATEGORY_MAP: dict[str, list[str]] = {
-    "main": ["MAINLINE"],
-    "activities": ["ACTIVITY", "MINI_ACTIVITY"],
-    "memoirs": ["NONE"],
-}
-
-
-def _story_zip_path(story_key: str) -> str:
-    return f"zh_CN/gamedata/story/{story_key}.json"
-
-
-# ---------------------------------------------------------------------------
-# Text cleaning
-# ---------------------------------------------------------------------------
-
-_RICH_TAG_RE = re.compile(r"<[^>]+>")
-_MEMOIR_EVENT_RE = re.compile(r"^story_([a-z0-9_]+)_set_\d+$")
-
-
-def _clean_text(text: str) -> str:
-    """Remove rich-text tags and replace {@nickname} with 博士."""
-    text = text.replace("{@nickname}", "博士")
-    text = _RICH_TAG_RE.sub("", text)
-    return text.strip()
-
-
-# ---------------------------------------------------------------------------
-# Data types
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class StoryLine:
-    type: Literal["dialog", "narration", "choice"]
-    role: str | None   # speaker name; None for narration/choice
-    text: str
-
-
-@dataclass(frozen=True)
-class StoryChapter:
-    story_key: str
-    story_code: str
-    story_name: str
-    avg_tag: str | None
-    event_name: str
-    story_info: str
-    lines: list[StoryLine]
-
-
-@dataclass(frozen=True)
-class EventInfo:
-    event_id: str
-    name: str
-    entry_type: str
-    story_count: int
-
-
-@dataclass(frozen=True)
-class ChapterSummary:
-    story_key: str
-    story_code: str
-    story_name: str
-    avg_tag: str | None
-    sort_order: int
-
-
-@dataclass(frozen=True)
-class ActivityResult:
-    event_id: str
-    event_name: str
-    total_chapters: int
-    has_more: bool
-    chapters: list[StoryChapter]
-
-
-@dataclass(frozen=True)
-class MemoirChapter:
-    event_id: str
-    story_key: str
-    story_code: str
-    story_name: str
-
-
-@dataclass(frozen=True)
-class OperatorMemoirResult:
-    operator_name: str
-    internal_code: str
-    operator_id: str
-    total_chapters: int
-    chapters: list[MemoirChapter]
-
-
-@dataclass(frozen=True)
-class _StorySearchChapter:
-    event_id: str
-    story_code: str
-    lines: tuple[StoryLine, ...]
-
-
-@dataclass(frozen=True)
-class _StorySearchRecord:
-    chapter_index: int
-    line_index: int
-    line: StoryLine
-
-
-@dataclass(frozen=True)
-class _StorySearchIndex:
-    event_ids: frozenset[str]
-    chapters: tuple[_StorySearchChapter, ...]
-    records: tuple[_StorySearchRecord, ...]
+from prts_mcp.data.story_reader import (
+    ActivityResult,
+    ChapterSummary,
+    EventInfo,
+    MemoirChapter,
+    OperatorMemoirResult,
+    StoryChapter,
+    StoryLine,
+    _clean_text,
+    _parse_story_list,
+    list_story_events,
+    list_story_events_from_store,
+    list_stories,
+    list_stories_from_store,
+    read_activity,
+    read_activity_from_store,
+    read_story,
+    read_story_from_store,
+)
+from prts_mcp.data.story_search import (
+    search_stories,
+    search_stories_from_store,
+)
+from prts_mcp.data.story_memoir import (
+    get_operator_memoirs,
+    get_operator_memoirs_from_store,
+)
+from prts_mcp.data.story_summary import (
+    get_event_summary,
+    get_event_summary_from_store,
+    get_story_summary,
+    get_story_summary_from_store,
+)
+from prts_mcp.data.story_search import clear_search_cache as _clear_search_cache
+from prts_mcp.data.story_memoir import clear_chardict_cache as _clear_chardict_cache
 
 
 def clear_story_caches() -> None:
     """Clear cached story search indexes after story data changes."""
-    _cached_story_search_index.cache_clear()
-    _cached_chardict.cache_clear()
-
-
-# ---------------------------------------------------------------------------
-# Core helpers
-# ---------------------------------------------------------------------------
-
-
-def _parse_story_list(story_list: list[dict]) -> list[StoryLine]:
-    """Convert raw storyList entries into cleaned StoryLine objects."""
-    lines: list[StoryLine] = []
-    for item in story_list:
-        prop = item.get("prop", "")
-        attrs = item.get("attributes", {})
-
-        prop_lower = prop.lower()
-        if prop_lower == "name":
-            name = attrs.get("name") or ""
-            content = attrs.get("content") or ""
-            if content:
-                lines.append(StoryLine(
-                    type="dialog",
-                    role=_clean_text(name) if name else None,
-                    text=_clean_text(content),
-                ))
-        elif prop_lower in ("sticker", "subtitle", "animtext"):
-            content = attrs.get("content") or attrs.get("text") or ""
-            if content:
-                lines.append(StoryLine(type="narration", role=None, text=_clean_text(content)))
-        elif prop_lower == "decision":
-            options = attrs.get("options") or []
-            for opt in options:
-                # options elements may be plain strings or dicts with a "text" key
-                text = opt if isinstance(opt, str) else (opt.get("text") or "")
-                if text:
-                    lines.append(StoryLine(type="choice", role=None, text=_clean_text(str(text))))
-
-    return lines
-
-
-def _story_store(zip_path: Path) -> ZipStore:
-    return ZipStore(zip_path)
-
-
-def _load_json(store: JsonStore, path: str) -> dict | list:
-    return store.read_json(path)
-
-
-# ---------------------------------------------------------------------------
-# Public API
-# ---------------------------------------------------------------------------
-
-
-def list_story_events(
-    zip_path: Path,
-    category: str | None = None,
-) -> list[EventInfo]:
-    """Return a list of events from story_review_table.json.
-
-    Args:
-        zip_path: Path to zh_CN.zip.
-        category: Optional filter — "main" or "activities".
-                  If None, all events are returned.
-    """
-    allowed_types: list[str] | None = _CATEGORY_MAP.get(category) if category else None
-
-    with _story_store(zip_path) as store:
-        return list_story_events_from_store(store, category=category)
-
-
-def list_story_events_from_store(
-    store: JsonStore,
-    category: str | None = None,
-) -> list[EventInfo]:
-    """Return a list of events from story_review_table.json using a JSON store."""
-    allowed_types: list[str] | None = _CATEGORY_MAP.get(category) if category else None
-    table: dict = _load_json(store, _STORY_REVIEW_TABLE)  # type: ignore[assignment]
-
-    events = []
-    for event_id, entry in table.items():
-        entry_type = entry.get("entryType", "NONE")
-        if allowed_types is not None:
-            if allowed_types == ["NONE"]:
-                # "memoirs" category: NONE entries whose event_id matches memoir pattern
-                if not _is_memoir_event(event_id):
-                    continue
-            elif entry_type not in allowed_types:
-                continue
-        story_count = len(entry.get("infoUnlockDatas") or [])
-        events.append(EventInfo(
-            event_id=event_id,
-            name=entry.get("name") or event_id,
-            entry_type=entry_type,
-            story_count=story_count,
-        ))
-
-    return events
-
-
-def list_stories(zip_path: Path, event_id: str) -> list[ChapterSummary]:
-    """Return ordered chapter list for an event.
-
-    Args:
-        zip_path: Path to zh_CN.zip.
-        event_id: Event key, e.g. "act31side".
-
-    Returns:
-        Chapters sorted by storySort.
-
-    Raises:
-        KeyError: If event_id is not found in story_review_table.
-    """
-    with _story_store(zip_path) as store:
-        return list_stories_from_store(store, event_id)
-
-
-def list_stories_from_store(store: JsonStore, event_id: str) -> list[ChapterSummary]:
-    """Return ordered chapter list for an event using a JSON store."""
-    table: dict = _load_json(store, _STORY_REVIEW_TABLE)  # type: ignore[assignment]
-
-    entry = table.get(event_id)
-    if entry is None:
-        raise KeyError(f"Event not found: {event_id!r}")
-
-    chapters = []
-    for d in sorted(entry.get("infoUnlockDatas") or [], key=lambda x: x.get("storySort", 0)):
-        story_key = d.get("storyTxt")
-        if not story_key:
-            continue
-        chapters.append(ChapterSummary(
-            story_key=story_key,
-            story_code=d.get("storyCode") or "",
-            story_name=d.get("storyName") or "",
-            avg_tag=d.get("avgTag"),
-            sort_order=d.get("storySort", 0),
-        ))
-
-    return chapters
-
-
-def read_story(
-    zip_path: Path,
-    story_key: str,
-    include_narration: bool = True,
-) -> StoryChapter:
-    """Read and parse a single story chapter.
-
-    Args:
-        zip_path: Path to zh_CN.zip.
-        story_key: Story key from storyTxt / storyinfo.json, e.g.
-                   "activities/act31side/level_act31side_01_beg".
-        include_narration: Whether to include narration/scene lines.
-
-    Raises:
-        KeyError: If the story file is not found in the zip.
-    """
-    with _story_store(zip_path) as store:
-        return read_story_from_store(store, story_key, include_narration=include_narration)
-
-
-def read_story_from_store(
-    store: JsonStore,
-    story_key: str,
-    include_narration: bool = True,
-) -> StoryChapter:
-    """Read and parse a single story chapter using a JSON store."""
-    story_path = _story_zip_path(story_key)
-    if not store.exists(story_path):
-        raise KeyError(f"Story not found in store: {story_key!r}")
-    raw: dict = _load_json(store, story_path)  # type: ignore[assignment]
-
-    all_lines = _parse_story_list(raw.get("storyList") or [])
-    if not include_narration:
-        all_lines = [
-            ln for ln in all_lines
-            if ln.type != "narration"
-            and not (ln.type == "dialog" and ln.role is None)
-        ]
-
-    return StoryChapter(
-        story_key=story_key,
-        story_code=raw.get("storyCode") or "",
-        story_name=raw.get("storyName") or "",
-        avg_tag=raw.get("avgTag"),
-        event_name=raw.get("eventName") or "",
-        story_info=raw.get("storyInfo") or "",
-        lines=all_lines,
-    )
-
-
-def read_activity(
-    zip_path: Path,
-    event_id: str,
-    include_narration: bool = True,
-    page: int | None = None,
-    page_size: int = 5,
-) -> ActivityResult:
-    """Read all chapters of an activity in official story order.
-
-    Args:
-        zip_path: Path to zh_CN.zip.
-        event_id: Event key, e.g. "act31side".
-        include_narration: Whether to include narration lines.
-        page: 1-based page index. None returns all chapters.
-        page_size: Chapters per page (used only when page is set).
-
-    Returns:
-        ActivityResult with event metadata and (paginated) chapters.
-
-    Raises:
-        KeyError: If event_id is not found.
-    """
-    with _story_store(zip_path) as store:
-        return read_activity_from_store(
-            store,
-            event_id,
-            include_narration=include_narration,
-            page=page,
-            page_size=page_size,
-        )
-
-
-def read_activity_from_store(
-    store: JsonStore,
-    event_id: str,
-    include_narration: bool = True,
-    page: int | None = None,
-    page_size: int = 5,
-) -> ActivityResult:
-    """Read all chapters of an activity in official story order using a JSON store."""
-    summaries = list_stories_from_store(store, event_id)
-    total = len(summaries)
-
-    if page is not None:
-        if page < 1:
-            raise ValueError("page 参数必须 >= 1")
-        start = (page - 1) * page_size
-        end = start + page_size
-        selected = summaries[start:end]
-        has_more = end < total
-    else:
-        selected = summaries
-        has_more = False
-
-    chapters = []
-    event_name = ""
-    for summary in selected:
-        try:
-            chapter = read_story_from_store(store, summary.story_key, include_narration)
-            if not event_name:
-                event_name = chapter.event_name
-            chapters.append(chapter)
-        except (KeyError, FileNotFoundError, json.JSONDecodeError):
-            # Story file missing or corrupt — skip silently
-            pass
-
-    return ActivityResult(
-        event_id=event_id,
-        event_name=event_name,
-        total_chapters=total,
-        has_more=has_more,
-        chapters=chapters,
-    )
-
-
-def _load_chardict(store: JsonStore) -> dict[str, dict[str, object]]:
-    """Load chardict.json from the story store."""
-    descriptor = _chardict_store_descriptor(store)
-    if descriptor is not None:
-        return _cached_chardict(descriptor)
-    if not store.exists(_CHARDICT):
-        return {}
-    return _read_chardict_from_store(store)
-
-
-def _read_chardict_from_store(store: JsonStore) -> dict[str, dict[str, object]]:
-    try:
-        data: dict = _load_json(store, _CHARDICT)  # type: ignore[assignment]
-        return {str(k): v for k, v in data.items() if isinstance(v, dict)}
-    except (KeyError, FileNotFoundError, json.JSONDecodeError):
-        return {}
-
-
-def _chardict_store_descriptor(store: JsonStore) -> tuple[str, str, int, int] | None:
-    if isinstance(store, ZipStore):
-        path = store.zip_path.resolve()
-        if not path.is_file():
-            return None
-        stat = path.stat()
-        return ("zip", str(path), stat.st_size, stat.st_mtime_ns)
-    if isinstance(store, DirectoryStore):
-        root = store.root.resolve()
-        chardict = root / _CHARDICT
-        if not chardict.is_file():
-            return None
-        stat = chardict.stat()
-        return ("directory", str(root), stat.st_size, stat.st_mtime_ns)
-    return None
-
-
-@lru_cache(maxsize=4)
-def _cached_chardict(
-    descriptor: tuple[str, str, int, int],
-) -> dict[str, dict[str, object]]:
-    kind, source, _size, _mtime_ns = descriptor
-    store: JsonStore
-    if kind == "zip":
-        store = ZipStore(source)
-    else:
-        store = DirectoryStore(source)
-    try:
-        return _read_chardict_from_store(store)
-    finally:
-        store.close()
-
-
-def _resolve_operator_code(
-    chardict: dict[str, dict[str, object]],
-    operator_name: str,
-) -> str | None:
-    """Reverse lookup: operator display name -> internal code (e.g. '能天使' -> 'angel')."""
-    for code, info in chardict.items():
-        if info.get("name") == operator_name:
-            return code
-    return None
-
-
-def _is_memoir_event(event_id: str) -> bool:
-    """Check if an event_id matches the memoir pattern story_{code}_set_N."""
-    return bool(_MEMOIR_EVENT_RE.match(event_id))
-
-
-def get_operator_memoirs(
-    zip_path: Path,
-    operator_name: str,
-) -> OperatorMemoirResult:
-    """Return all memoir chapters for an operator by Chinese name.
-
-    Raises:
-        KeyError: If the operator is not found in chardict or has no memoirs.
-    """
-    with _story_store(zip_path) as store:
-        return get_operator_memoirs_from_store(store, operator_name)
-
-
-def get_operator_memoirs_from_store(
-    store: JsonStore,
-    operator_name: str,
-) -> OperatorMemoirResult:
-    """Return all memoir chapters for an operator by Chinese name using a JSON store."""
-    chardict = _load_chardict(store)
-    if not chardict:
-        raise KeyError("chardict.json 未在 story zip 中找到。")
-
-    code = _resolve_operator_code(chardict, operator_name)
-    if code is None:
-        raise KeyError(
-            f"未找到干员名称 '{operator_name}' 对应的内部代码。请使用游戏内中文名称。"
-        )
-
-    char_info = chardict[code]
-    operator_id = char_info.get("id", "")
-
-    table: dict = _load_json(store, _STORY_REVIEW_TABLE)  # type: ignore[assignment]
-    prefix = f"story_{code}_set_"
-    memoir_events = sorted(
-        (eid for eid in table if eid.startswith(prefix)),
-        key=lambda eid: int(eid.rsplit("_", 1)[-1]) if eid.rsplit("_", 1)[-1].isdigit() else 0,
-    )
-
-    if not memoir_events:
-        raise KeyError(f"干员 '{operator_name}' (code={code}) 暂无密录数据。")
-
-    chapters: list[MemoirChapter] = []
-    for event_id in memoir_events:
-        entry = table[event_id]
-        datas = sorted(
-            entry.get("infoUnlockDatas") or [],
-            key=lambda x: x.get("storySort", 0),
-        )
-        for d in datas:
-            story_key = d.get("storyTxt")
-            if not story_key:
-                continue
-            chapters.append(MemoirChapter(
-                event_id=event_id,
-                story_key=story_key,
-                story_code=d.get("storyCode", ""),
-                story_name=d.get("storyName", ""),
-            ))
-
-    return OperatorMemoirResult(
-        operator_name=operator_name,
-        internal_code=code,
-        operator_id=operator_id,
-        total_chapters=len(chapters),
-        chapters=chapters,
-    )
-
-
-# ---------------------------------------------------------------------------
-# Search helpers
-# ---------------------------------------------------------------------------
-
-def _format_story_line(line: StoryLine) -> str:
-    """Format a single StoryLine for display in search results."""
-    if line.type == "dialog":
-        return f"{line.role or '（旁白）'}：{line.text}"
-    elif line.type == "narration":
-        return f"*{line.text}*"
-    else:
-        return f"【选项】{line.text}"
-
-
-_VALID_LINE_TYPES = {"dialog", "narration", "choice"}
-
-
-def search_stories(
-    zip_path: Path,
-    pattern: str,
-    character: str | None = None,
-    line_type: str | None = None,
-    context_lines: int = 1,
-    max_results: int = 30,
-    event_id: str | None = None,
-) -> str:
-    """Search story text across all events (or a single event).
-
-    Convenience wrapper around search_stories_from_store that auto-creates
-    a ZipStore from *zip_path*.
-    """
-    with _story_store(zip_path) as store:
-        return search_stories_from_store(
-            store,
-            pattern,
-            character=character,
-            line_type=line_type,
-            context_lines=context_lines,
-            max_results=max_results,
-            event_id=event_id,
-        )
-
-
-def search_stories_from_store(
-    store: JsonStore,
-    pattern: str,
-    character: str | None = None,
-    line_type: str | None = None,
-    context_lines: int = 1,
-    max_results: int = 30,
-    event_id: str | None = None,
-) -> str:
-    """Full-text search across story dialogue, narration and choice lines.
-
-    Args:
-        store: JsonStore (ZipStore or DirectoryStore) for story data.
-        pattern: Case-insensitive regex pattern.
-        character: Optional speaker name filter (dialog lines only).
-        line_type: Optional line type filter — "dialog", "narration" or "choice".
-        context_lines: Number of surrounding lines to include per match.
-        max_results: Maximum matches to return.
-        event_id: Optional event filter (e.g. "act31side").
-
-    Returns:
-        Formatted multi-block search result string.
-    """
-    if max_results < 1:
-        return "max_results 必须 >= 1。"
-    if max_results > 100:
-        return "max_results 必须 <= 100。"
-    if context_lines < 0:
-        return "context_lines 必须 >= 0。"
-    if context_lines > 5:
-        return "context_lines 必须 <= 5。"
-
-    try:
-        regex = re.compile(pattern, re.IGNORECASE)
-    except re.error as exc:
-        return f"正则表达式无效：{exc}"
-
-    if line_type is not None and line_type not in _VALID_LINE_TYPES:
-        return f"无效的 line_type：{line_type!r}，可选值：{', '.join(sorted(_VALID_LINE_TYPES))}"
-
-    try:
-        index = _story_search_index(store)
-    except Exception as exc:
-        return f"读取剧情数据索引失败：{exc}"
-
-    if event_id is not None and event_id not in index.event_ids:
-        return f"未找到匹配的活动：{event_id!r}。"
-
-    results: list[dict] = []
-
-    for record in index.records:
-        if len(results) >= max_results:
-            break
-
-        chapter = index.chapters[record.chapter_index]
-        line = record.line
-        if event_id is not None and chapter.event_id != event_id:
-            continue
-        if character is not None:
-            if line.type != "dialog" or (line.role or "").lower() != character.lower():
-                continue
-        if line_type is not None and line.type != line_type:
-            continue
-        if not regex.search(line.text):
-            continue
-
-        start = max(0, record.line_index - context_lines)
-        end = min(len(chapter.lines), record.line_index + context_lines + 1)
-        ctx_parts: list[str] = []
-        for j in range(start, end):
-            prefix = ">>> " if j == record.line_index else "    "
-            ctx_parts.append(prefix + _format_story_line(chapter.lines[j]))
-
-        results.append({
-            "event_id": chapter.event_id,
-            "story_code": chapter.story_code,
-            "line_number": record.line_index + 1,
-            "context": "\n".join(ctx_parts),
-        })
-
-    if not results:
-        filter_desc = "。".join(
-            f for f in [
-                f"event_id={event_id!r}" if event_id else "",
-                f"character={character!r}" if character else "",
-                f"line_type={line_type!r}" if line_type else "",
-            ] if f
-        )
-        filter_suffix = f"（过滤条件：{filter_desc}）" if filter_desc else ""
-        return f"未找到匹配 '{pattern}' 的剧情台词。{filter_suffix}"
-
-    parts = [f"# 搜索 \"{pattern}\" 的结果（共 {len(results)} 条）"]
-    for r in results:
-        parts.append(
-            f"\n---\n\n"
-            f"[stories/{r['event_id']}/{r['story_code']} L{r['line_number']}]\n"
-            f"{r['context']}"
-        )
-
-    return "\n".join(parts)
-
-
-def _story_search_index(store: JsonStore) -> _StorySearchIndex:
-    descriptor = _story_store_descriptor(store)
-    if descriptor is None:
-        return _build_story_search_index(store)
-    return _cached_story_search_index(descriptor)
-
-
-def _story_store_descriptor(store: JsonStore) -> tuple[str, str, int, int] | None:
-    if isinstance(store, ZipStore):
-        path = store.zip_path.resolve()
-        if not path.is_file():
-            return None
-        stat = path.stat()
-        return ("zip", str(path), stat.st_size, stat.st_mtime_ns)
-    if isinstance(store, DirectoryStore):
-        root = store.root.resolve()
-        review = root / _STORY_REVIEW_TABLE
-        if not review.is_file():
-            return None
-        stat = review.stat()
-        return ("directory", str(root), stat.st_size, stat.st_mtime_ns)
-    return None
-
-
-@lru_cache(maxsize=4)
-def _cached_story_search_index(
-    descriptor: tuple[str, str, int, int],
-) -> _StorySearchIndex:
-    kind, source, _size, _mtime_ns = descriptor
-    store: JsonStore
-    if kind == "zip":
-        store = ZipStore(source)
-    else:
-        store = DirectoryStore(source)
-    try:
-        return _build_story_search_index(store)
-    finally:
-        store.close()
-
-
-def _build_story_search_index(store: JsonStore) -> _StorySearchIndex:
-    table: dict = _load_json(store, _STORY_REVIEW_TABLE)  # type: ignore[assignment]
-    event_ids: set[str] = set()
-    chapters: list[_StorySearchChapter] = []
-    records: list[_StorySearchRecord] = []
-
-    for ev_id, entry in table.items():
-        datas = sorted(
-            entry.get("infoUnlockDatas") or [],
-            key=lambda x: x.get("storySort", 0),
-        )
-        if not datas:
-            continue
-        # Only include NONE entries that are operator memoirs
-        if entry.get("entryType", "NONE") == "NONE" and not _is_memoir_event(ev_id):
-            continue
-        event_ids.add(ev_id)
-        for d in datas:
-            story_key = d.get("storyTxt")
-            if not story_key:
-                continue
-            try:
-                chapter = read_story_from_store(store, story_key, include_narration=True)
-            except (KeyError, FileNotFoundError, json.JSONDecodeError):
-                continue
-            chapter_index = len(chapters)
-            indexed_chapter = _StorySearchChapter(
-                event_id=ev_id,
-                story_code=d.get("storyCode", ""),
-                lines=tuple(chapter.lines),
-            )
-            chapters.append(indexed_chapter)
-            records.extend(
-                _StorySearchRecord(chapter_index=chapter_index, line_index=i, line=line)
-                for i, line in enumerate(indexed_chapter.lines)
-            )
-
-    return _StorySearchIndex(
-        event_ids=frozenset(event_ids),
-        chapters=tuple(chapters),
-        records=tuple(records),
-    )
-
-
-# ---------------------------------------------------------------------------
-# Event summary
-# ---------------------------------------------------------------------------
-
-
-def get_event_summary(zip_path: Path, event_id: str) -> str:
-    """Return a chapter-by-chapter summary overview of an event.
-
-    Convenience wrapper around get_event_summary_from_store.
-    """
-    with _story_store(zip_path) as store:
-        return get_event_summary_from_store(store, event_id)
-
-
-def get_event_summary_from_store(store: JsonStore, event_id: str) -> str:
-    """Return a chapter-by-chapter summary overview of an event.
-
-    Reads story_review_table for chapter ordering and storyinfo.json
-    for per-chapter summaries, producing a narrative table of contents
-    suitable for getting the big picture of an event at a glance.
-    """
-    # --- event metadata ---
-    try:
-        table: dict = _load_json(store, _STORY_REVIEW_TABLE)
-    except Exception as exc:
-        return f"读取剧情数据索引失败：{exc}"
-
-    entry = table.get(event_id)
-    if entry is None:
-        return f"未找到活动：{event_id!r}。请先调用 list_story_events 确认活动 ID。"
-
-    event_name = entry.get("name") or event_id
-    datas = sorted(
-        entry.get("infoUnlockDatas") or [],
-        key=lambda x: x.get("storySort", 0),
-    )
-
-    if not datas:
-        return f"活动 {event_id!r}（{event_name}）暂无剧情章节。"
-
-    # --- load summary index ---
-    summaries: dict[str, str] = {}
-    if store.exists(_STORYINFO):
-        try:
-            raw = _load_json(store, _STORYINFO)
-            if isinstance(raw, dict):
-                summaries = {str(k): str(v) for k, v in raw.items() if v}
-        except Exception:
-            pass  # storyinfo.json is optional for this tool
-
-    # --- tier 1: LLM event summary ---
-    event_summary_text = ""
-    if store.exists(_EVENT_SUMMARIES):
-        try:
-            raw = _load_json(store, _EVENT_SUMMARIES)
-            if isinstance(raw, dict):
-                event_summary_text = str(raw.get(event_id) or "").strip()
-        except Exception:
-            pass
-
-    # --- build output ---
-    total = len(datas)
-    lines = [f"# {event_name} — 共 {total} 章"]
-    if event_summary_text:
-        lines.append(f"\n{event_summary_text}")
-    for d in datas:
-        story_key = d.get("storyTxt")
-        if not story_key:
-            continue
-        code = d.get("storyCode", "")
-        name = d.get("storyName", "")
-        tag = f"[{d['avgTag']}] " if d.get("avgTag") else ""
-
-        summary = summaries.get(story_key, "")
-        if summary:
-            lines.append(f"\n{code} {tag}{name}\n  {summary}")
-        else:
-            lines.append(f"\n{code} {tag}{name}")
-
-    return "\n".join(lines)
-
-
-# ---------------------------------------------------------------------------
-# Per-chapter summary
-# ---------------------------------------------------------------------------
-
-
-def get_story_summary(zip_path: Path, story_key: str) -> str:
-    """Return a summary for a single story chapter.
-
-    Convenience wrapper around get_story_summary_from_store.
-    """
-    with _story_store(zip_path) as store:
-        return get_story_summary_from_store(store, story_key)
-
-
-def get_story_summary_from_store(store: JsonStore, story_key: str) -> str:
-    """Return a summary for a single story chapter.
-
-    Fallback chain:
-    1. zh_CN/summaries.json — LLM-generated long summary (future)
-    2. zh_CN/storyinfo.json — official one-line summary
-    3. Chapter JSON ``storyInfo`` field — identical to #2, last resort
-    """
-    # --- tier 1: LLM summaries (future) ---
-    if store.exists(_SUMMARIES):
-        try:
-            raw = _load_json(store, _SUMMARIES)
-            if isinstance(raw, dict):
-                text = raw.get(story_key)
-                if text and isinstance(text, str):
-                    return text.strip()
-        except Exception:
-            pass
-
-    # --- tier 2: storyinfo.json ---
-    if store.exists(_STORYINFO):
-        try:
-            raw = _load_json(store, _STORYINFO)
-            if isinstance(raw, dict):
-                text = raw.get(story_key)
-                if text and isinstance(text, str):
-                    return text.strip()
-        except Exception:
-            pass
-
-    # --- tier 3: chapter JSON storyInfo ---
-    story_path = _story_zip_path(story_key)
-    if store.exists(story_path):
-        try:
-            raw = _load_json(store, story_path)
-            if isinstance(raw, dict):
-                text = raw.get("storyInfo")
-                if text and isinstance(text, str):
-                    return text.strip()
-        except Exception:
-            pass
-
-    return f"未找到剧情章节 '{story_key}' 的梗概。"
+    _clear_search_cache()
+    _clear_chardict_cache()
+
+
+__all__ = [
+    # Types
+    "StoryLine",
+    "StoryChapter",
+    "EventInfo",
+    "ChapterSummary",
+    "ActivityResult",
+    "MemoirChapter",
+    "OperatorMemoirResult",
+    # Reader
+    "list_story_events",
+    "list_story_events_from_store",
+    "list_stories",
+    "list_stories_from_store",
+    "read_story",
+    "read_story_from_store",
+    "read_activity",
+    "read_activity_from_store",
+    # Search
+    "search_stories",
+    "search_stories_from_store",
+    # Memoir
+    "get_operator_memoirs",
+    "get_operator_memoirs_from_store",
+    # Summary
+    "get_event_summary",
+    "get_event_summary_from_store",
+    "get_story_summary",
+    "get_story_summary_from_store",
+    # Cache management
+    "clear_story_caches",
+]
